Week 6 Weekend Assignment: CRUD semi-Operational Broke the edit button.... working on that.
TODO:
Edit & Delete operation Needs to be user specific

+++ src/main/java/com/crooks/controllers/LunchController.java
@@ -20,13 +20,14 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import javax.servlet.http.HttpSession;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
 
 /**
  * Created by johncrooks on 6/23/16.
  */
 @Controller
 public class LunchController {
-    Boolean openEdit = false;
+
 
     @Autowired
     UserRepository userRepo;
@@ -36,12 +37,14 @@ public class LunchController {
     @RequestMapping(path="/", method = RequestMethod.GET)
     public String home(HttpSession session, Model model){
         String username = (String) session.getAttribute("username");
+        User user = userRepo.findByName(username);
 
         model.addAttribute("username", username);
         model.addAttribute("lunches", lunchRepo.findAll());
         model.addAttribute("now", LocalDate.now());
-        model.addAttribute("openEdit", openEdit);
-
+        if (user !=null) {
+            model.addAttribute("openEdit", user.getOpenEdit());
+        }
         return "home";
     }
 
@@ -67,7 +70,12 @@ public class LunchController {
             return "redirect:/";
         }
         Lunch lunch = new Lunch(LocalDate.parse(date),restaurant, price,description, user);
+        List<Lunch> tempLunch = user.getLunchList();
+        tempLunch.add(lunch);
+        user.setLunchList(tempLunch);
+        user.setOpenEdit(false);
         lunchRepo.save(lunch);
+
 
         return "redirect:/";
 
@@ -76,29 +84,45 @@ public class LunchController {
     public String editLunch(HttpSession session, int id, String date, String restaurant, String description, double price){
         String username = (String) session.getAttribute("username");
         User user = userRepo.findByName(username);
-
-        Lunch lunch = new Lunch(id, LocalDate.parse(date),restaurant, price,description, user);
-        lunchRepo.save(lunch);
-
-        openEdit = false;
+        if(username==null){
+            return "redirect:/";
+        }else if( username.equals(user.getName())){
+            Lunch lunch = new Lunch(id, LocalDate.parse(date), restaurant, price, description, user);
+            lunchRepo.save(lunch);
+            user.setOpenEdit(false);
+        }
         return "redirect:/";
 
     }
 
     @RequestMapping(value="/openEdit", method = RequestMethod.POST)
-    public String openEdit(){
-        if(openEdit){
-           openEdit= !openEdit;
-        }else if (!openEdit){
-            openEdit = !openEdit;
+    public String openEdit(HttpSession session, String name) throws Exception {
+        String username = (String) session.getAttribute("username");
+        User user1 = userRepo.findByName(username);
+        if (name.equals(user1.getName())){
+            if(user1.getOpenEdit()){
+                user1.setOpenEdit(!user1.getOpenEdit());
+            }else if (!user1.getOpenEdit()){
+                user1.setOpenEdit(!user1.getOpenEdit());
+            }
+        } else{
+            throw new Exception("This isn't your post, You can't edit this...");
         }
-        return "redirect:/";
+
+            return "redirect:/";
     }
 
     @RequestMapping(path="/delete-lunch", method=RequestMethod.POST)
-    public String deleteLunch(int id){
-        lunchRepo.delete(id);
+    public String deleteLunch(HttpSession session, int id) throws Exception {
+        String username = (String) session.getAttribute("username");
+        User user = userRepo.findByName(username);
 
+        if (user.getId()==lunchRepo.findOne(id).getUser().getId()) {
+
+            lunchRepo.delete(id);
+        } else{
+            throw new Exception("---- WHOA WHOA WHOA, this isn't yours to delete! ----");
+        }
         return "redirect:/";
     }
